% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{duke_palettes}
\alias{duke_palettes}
\title{Complete list of palettes}
\format{An object of class \code{list} of length 4.}
\usage{
duke_palettes
}
\description{
Use \code{\link{duke_palette}} to construct palettes of desired length.
}
\keyword{datasets}
