% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dukecolor.R
\docType{package}
\name{dukecolor}
\alias{dukecolor}
\alias{dukecolor-package}
\title{dukecolor}
\description{
list of palettes from Duke University Style guide
}
\details{
list of palettes from Duke University Style guide
}
